# Reduce the number of calls to hash.update