--- a/lib/Chunk.js
+++ b/lib/Chunk.js
@@ -349,11 +349,14 @@ class Chunk {
 		return this._modules.size === 0;
 	}
 
-	updateHash(hash) {
-		hash.update(`${this.id} `);
-		hash.update(this.ids ? this.ids.join(",") : "");
-		hash.update(`${this.name || ""} `);
-		this._modules.forEach(m => m.updateHash(hash));
+	hashPart() {
+		let str = `${this.id} `;
+		str += this.ids ? this.ids.join(",") : "";
+		str += `${this.name || ""} `;
+		this._modules.forEach(m => {
+			str += m.hashPart();
+		});
+		return str;
 	}
 
 	canBeIntegrated(otherChunk) {