--- a/lib/ChunkTemplate.js
+++ b/lib/ChunkTemplate.js
@@ -24,8 +24,7 @@ module.exports = class ChunkTemplate extends Template {
 	}
 
 	updateHash(hash) {
-		hash.update("ChunkTemplate");
-		hash.update("2");
+		hash.update("ChunkTemplate2");
 		this.applyPlugins("hash", hash);
 	}