--- a/lib/Compilation.js
+++ b/lib/Compilation.js
@@ -1328,10 +1328,14 @@ class Compilation extends Tapable {
 		});
 		for(let i = 0; i < chunks.length; i++) {
 			const chunk = chunks[i];
-			const chunkHash = crypto.createHash(hashFunction);
+			let strToHash = "";
 			if(outputOptions.hashSalt)
-				chunkHash.update(outputOptions.hashSalt);
-			chunk.updateHash(chunkHash);
+				strToHash += outputOptions.hashSalt;
+			strToHash += chunk.hashPart();
+
+			const chunkHash = crypto.createHash(hashFunction);
+			chunkHash.update(strToHash); // TODO avoid more calls to hash.update
+
 			if(chunk.hasRuntime()) {
 				this.mainTemplate.updateHashForChunk(chunkHash, chunk);
 			} else {