--- a/lib/DependenciesBlock.js
+++ b/lib/DependenciesBlock.js
@@ -39,14 +39,18 @@ class DependenciesBlock {
 		this.dependencies.push(dependency);
 	}
 
-	updateHash(hash) {
-		function updateHash(i) {
-			i.updateHash(hash);
+	hashPart() {
+		let str = "";
+
+		function addHashPartToStr(i) {
+			str += i.hashPart();
 		}
 
-		this.dependencies.forEach(updateHash);
-		this.blocks.forEach(updateHash);
-		this.variables.forEach(updateHash);
+		this.dependencies.forEach(addHashPartToStr);
+		this.blocks.forEach(addHashPartToStr);
+		this.variables.forEach(addHashPartToStr);
+
+		return str;
 	}
 
 	disconnect() {