--- a/lib/DependenciesBlockVariable.js
+++ b/lib/DependenciesBlockVariable.js
@@ -14,12 +14,14 @@ class DependenciesBlockVariable {
 		this.dependencies = dependencies || [];
 	}
 
-	updateHash(hash) {
-		hash.update(this.name);
-		hash.update(this.expression);
+	hashPart() {
+		let dependenciesHashParts = "";
+
 		this.dependencies.forEach(d => {
-			d.updateHash(hash);
+			dependenciesHashParts += d.hashPart();
 		});
+
+		return this.name + this.expression + dependenciesHashParts;
 	}
 
 	expressionSource(dependencyTemplates, outputOptions, requestShortener) {