--- a/lib/Dependency.js
+++ b/lib/Dependency.js
@@ -36,8 +36,8 @@ class Dependency {
 		return null;
 	}
 
-	updateHash(hash) {
-		hash.update((this.module && this.module.id) + "");
+	hashPart() {
+		return (this.module && this.module.id) + "";
 	}
 
 	disconnect() {