--- a/lib/MainTemplate.js
+++ b/lib/MainTemplate.js
@@ -213,9 +213,7 @@ module.exports = class MainTemplate extends Template {
 	}
 
 	updateHash(hash) {
-		hash.update("maintemplate");
-		hash.update("3");
-		hash.update(this.outputOptions.publicPath + "");
+		hash.update("maintemplate3" + this.outputOptions.publicPath);
 		this.applyPlugins("hash", hash);
 	}