--- a/lib/Module.js
+++ b/lib/Module.js
@@ -199,10 +199,8 @@ class Module extends DependenciesBlock {
 		return true;
 	}
 
-	updateHash(hash) {
-		hash.update(this.id + "" + this.used);
-		hash.update(JSON.stringify(this.usedExports));
-		super.updateHash(hash);
+	hashPart() {
+		return this.id + this.used + JSON.stringify(this.usedExports) + super.hashPart();
 	}
 
 	sortItems(sortChunks) {