--- a/lib/NormalModule.js
+++ b/lib/NormalModule.js
@@ -311,8 +311,7 @@ class NormalModule extends Module {
 	getHashDigest(dependencyTemplates) {
 		let dtHash = dependencyTemplatesHashMap.get("hash");
 		const hash = crypto.createHash("md5");
-		this.updateHash(hash);
-		hash.update(`${dtHash}`);
+		hash.update(`${this.hashPart()}${dtHash}`);
 		return hash.digest("hex");
 	}
 
@@ -531,24 +530,18 @@ class NormalModule extends Module {
 		return this._source ? this._source.size() : -1;
 	}
 
-	updateHashWithSource(hash) {
-		if(!this._source) {
-			hash.update("null");
-			return;
-		}
-		hash.update("source");
-		this._source.updateHash(hash);
+	sourceHashPart() {
+		return this._source
+			? "null"
+			: "source" + this._source.source();
 	}
 
-	updateHashWithMeta(hash) {
-		hash.update("meta");
-		hash.update(JSON.stringify(this.meta));
+	metaHashPart() {
+		return "meta" + JSON.stringify(this.meta);
 	}
 
-	updateHash(hash) {
-		this.updateHashWithSource(hash);
-		this.updateHashWithMeta(hash);
-		super.updateHash(hash);
+	hashPart() {
+		return this.sourceHashPart() + this.metaHashPart() + super.hashPart();
 	}
 
 }