--- a/lib/TemplatedPathPlugin.js
+++ b/lib/TemplatedPathPlugin.js
@@ -100,10 +100,13 @@ class TemplatedPathPlugin {
 			mainTemplate.plugin("hash-for-chunk", function(hash, chunk) {
 				const outputOptions = this.outputOptions;
 				const chunkFilename = outputOptions.chunkFilename || outputOptions.filename;
+				let strToHash = "";
 				if(REGEXP_CHUNKHASH_FOR_TEST.test(chunkFilename))
-					hash.update(JSON.stringify(chunk.getChunkMaps(true, true).hash));
+					strToHash += JSON.stringify(chunk.getChunkMaps(true, true).hash);
 				if(REGEXP_NAME_FOR_TEST.test(chunkFilename))
-					hash.update(JSON.stringify(chunk.getChunkMaps(true, true).name));
+					strToHash += JSON.stringify(chunk.getChunkMaps(true, true).name);
+				if(strToHash)
+					hash.update(strToHash);
 			});
 		});
 	}