I was profiling our webpack build on Node 8.9.0 and noticed that a fair
amount of time was spent calling hash.update. During the process of
building one of our bundles, ~115ms was spent inside of createHash. 95ms
of that time is spent in crypto's update itself. In the bottom up view
of the profiler, this is the 4th most expensive part of the profile, as
measured by self time.

It turns out that, concatenating up a string and calling crypto's update
once is much faster than calling update many times. In this commit, I
take a swing at replacing the pattern of passing around a hash object
that update is called on all the time, to concatenating up a string, and
reducing the number of calls to crypto's update.

After this change, my profiling shows that only 16.54ms is spent in
createHash. In the bottom up view, I verified that 8.5ms is now spent in
crypto's update, which moves it far from one of the most expensive
areas.

I think there are more opportunities like this, but I'm a little
hesitant to dig in much more at this point because of the staggering
amount of indirection which makes the code difficult to follow, the
profiling difficult to interpret, and the stack traces when I hit an
error not very helpful.

<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**What kind of change does this PR introduce?**

<!-- E.g. a bugfix, feature, refactoring, build related change, etc… -->
Optimization

**Did you add tests for your changes?**

<!-- Note that we won't merge your changes if you don't add tests -->
Not yet. I wanted to get feedback on this before I fixed the tests.

**If relevant, link to documentation update:**

<!-- Link PR from webpack/webpack.js.org here, or N/A -->

**Summary**

<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->
<!-- Try to link to an open issue for more information. -->
See above. https://github.com/webpack/webpack/issues/5718

**Does this PR introduce a breaking change?**

<!-- If this PR introduces a breaking change, please describe the impact and a migration path for existing applications. -->
I don't think so

**Other information**

Before:
<img width="1156" alt="screen shot 2017-12-05 at 2 19 01 pm" src="https://user-images.githubusercontent.com/195534/33633920-501cabbc-d9c7-11e7-800e-45522002e299.png">

After:
<img width="1056" alt="screen shot 2017-12-05 at 2 19 13 pm" src="https://user-images.githubusercontent.com/195534/33633921-50336f96-d9c7-11e7-8cdf-c0c4b478be0b.png">