@lencioni 2017-12-06
I added more of our bundles into the mix and profiled before and after this change again, and this seems to have a nice positive impact. Here are some flame charts.

Before
<img width="1246" alt="developer tools - node js 2017-12-05 16-11-56" src="https://user-images.githubusercontent.com/195534/33637663-2b62de3a-d9d7-11e7-9873-f1ac6f2328ff.png">

After
<img width="1242" alt="developer tools - node js 2017-12-05 16-13-20" src="https://user-images.githubusercontent.com/195534/33637677-45e8b720-d9d7-11e7-8fbd-4fc9da881d2e.png">

Zooming in a bit:

Before
<img width="907" alt="screen shot 2017-12-05 at 4 16 04 pm" src="https://user-images.githubusercontent.com/195534/33637765-ae8c20f0-d9d7-11e7-8741-7630105cecb3.png">

After
<img width="829" alt="screen shot 2017-12-05 at 4 16 16 pm" src="https://user-images.githubusercontent.com/195534/33637766-aea0d9e6-d9d7-11e7-918e-58984a233878.png">