@sokra 2017-12-06
This is similar to #6006 but more complex.

You could try to backport #6006 to webpack 3, profile it and send a PR.