@lencioni 2017-12-06
Ah thanks! I missed that one. Since v4 is so close, maybe I should start profiling on that version instead.